+++ QuickFIXn/Fields/Converters/DateTimeConverter.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Text;
 using System.Globalization;
 
 namespace QuickFix.Fields.Converters
@@ -61,16 +60,16 @@ namespace QuickFix.Fields.Converters
             System.DateTimeKind kind;
             int offset = 0;
 
-            if (dec.EndsWith("Z"))
+            if (dec.EndsWith('Z'))
             {
                 // UTC
                 dec = dec.Substring(0, dec.Length - 1);
                 kind = System.DateTimeKind.Utc;
             }
-            else if (dec.Contains("+") || dec.Contains("-"))
+            else if (dec.Contains('+') || dec.Contains('-'))
             {
                 // GMT offset
-                int n = dec.Contains("+") ? dec.IndexOf('+') : dec.IndexOf('-');
+                int n = dec.Contains('+') ? dec.IndexOf('+') : dec.IndexOf('-');
                 kind = System.DateTimeKind.Unspecified;
                 offset = int.Parse(dec.Substring(n));
                 dec = dec.Substring(0, n);